requirements.txt
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
streamlit>=1.28.0
matplotlib>=3.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
boto3>=1.28.0
cloudinary>=1.36.0

